// src/corporate-modules/corporate-bussiness-document/dtos/request/create.dto.ts
export class CreateCorporateBussinessDocumentDTO {
  company_register_document: string;
  company_name: string;
  address: string;
  sub_district: string;
  district: string;
  province: string;
  zip_code: string;
  country: string;
  is_VAT_registered: boolean;
  VAT_id: string;
  is_head_office: boolean;
  branch_no: string;
  VAT_address: string;
  VAT_certification: string;
}